+++ CHANGELOG.md
@@ -6,7 +6,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
 - Initial setup
 - Add Functions
   - Invoke-Idoit, Connect-Idoit, Disconnect-Idoit
@@ -16,3 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Get-IdoitVersion
   - Start-IdoitApiTrace, Stop-IdoitApiTrace
 
+- Add Functions
+  - New-IdoitObject, Set-IdoitCategory
+  - Get-Dialog, Get-LocationTree